styling moved, accordions fixed, tiny bug was caught, next steps are to standardize all the css within the new files (i.e color vars, padding etc)
