--- ui/src/App.css
@@ -1,355 +1,0 @@
-:root {
-  --header-height: 7em;
-  --footer-height: 7em;
-  --page-padding: 1.5em;
-  --fonts: "Helvetica Neue", "Helvetica", "Roboto", "Arial", sans‑serif;
-}
-
-#page {
-  min-height: calc(
-    100vh - var(--header-height) - var(--footer-height) - var(--page-padding) -
-      var(--page-padding)
-  );
-  margin: var(--page-padding) 0 3em 0;
-}
-
-.overviewText {
-  font-size: 16pt;
-}
-
-.pagination-container {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-}
-
-.accordion-button-group {
-  display: flex;
-  flex-direction: column;
-}
-
-.accordion {
-  height: auto !important;
-  margin-top: 0 !important;
-  margin-bottom: 1em !important;
-}
-
-.accordion-button-group .accordion-buttons-container {
-  position: absolute;
-  right: 1em;
-  margin: 2px 1em !important;
-}
-
-/* For nested accordion with depth of 1 */
-.accordion .accordion-button-group .accordion-buttons-container {
-  position: absolute;
-  right: 1em;
-  margin: 2px 2em !important;
-}
-
-/* 
- * For nested accordion with depth of 2 
- * (please for the love of all that is holy, do not have accordions nested more than 2 deep) 
- */
-.accordion .accordion .accordion-button-group .accordion-buttons-container {
-  position: absolute;
-  /* right: 1em;
-  margin: 2px 3em !important; */
-}
-
-.accordion-buttons-container > * {
-  margin: 0 0 0 0 !important;
-  border-radius: 0 !important;
-}
-
-.accordion-buttons-container :first-child {
-  margin: 0 0 0 0 !important;
-  /* border-top-left-radius: 0.28571429rem !important;
-  border-bottom-left-radius: 0.28571429rem !important; */
-}
-
-.accordion-buttons-container :last-child {
-  margin: 0 0 0 0 !important;
-  /* border-top-right-radius: 0.28571429rem !important;
-  border-bottom-right-radius: 0.28571429rem !important; */
-}
-
-.fake-a {
-  color: #1e70bf;
-  text-decoration: none;
-  cursor: pointer;
-}
-
-.spacer {
-  flex-grow: 1;
-}
-
-button.ui.button,
-.ui.icon.button,
-.ui.icon.buttons,
-.button,
-a.ui.button {
-  background-color: #f8f9fa;
-  color: #212529;
-  padding: 10px 20px;
-  margin-top: 0.5em;
-  margin-bottom: 0.5em;
-}
-
-*:not(i) {
-  font-family: var(--fonts) !important;
-}
-
-.pagination-container {
-  padding-top: 1em;
-}
-
-.clickable-student-name {
-  cursor: pointer;
-  color: #4183c4;
-  text-decoration: underline;
-}
-
-
-
-/* Dark Mode Styles */
-.dark-mode {
-  background-color: #121212;
-  color: #e0e0e0;
-}
-
-.dark-mode a {
-  color: #90caf9;
-}
-
-.dark-mode .fake-a {
-  color: #90CAF9;
-}
-
-.dark-mode #footer {
-  background-color: #1e1e1e;
-  color: #e0e0e0;
-}
-
-.dark-mode .ui.button {
-  background-color: #424242;
-  color: #fff;
-  border: 1px solid #fff;
-}
-
-.dark-mode .ui.button:hover {
-  background-color: #333;
-}
-
-body.dark-mode table, 
-body.dark-mode th, 
-body.dark-mode td {
-    border: 1px solid #444;
-}
-
-
-body.dark-mode .ui.header {
-  color: #ffffff;
-}
-
-body.dark-mode .clickable-student-name {
-  cursor: pointer;
-  color: #90CAF9;
-  text-decoration: underline;
-}
-
-.dark-mode .ui.segment,
-.dark-mode .ui.tab {
-    background-color: #222;
-    color: #fff;
-}
-
-.dark-mode .ui.menu {
-    background-color: black;
-    color: white;
-}
-
-.dark-mode .ui.tabular.menu .item {
-    color: white;
-    background-color: rgb(48, 48, 48);
-}
-
-.dark-mode .ui.tabular.menu .active.item {
-    background-color: #999999;
-    color: white;
-}
-
-.dark-mode .ui.pane {
-    background-color: #222;
-    color: #ddd;
-}
-
-.dark-mode .semester-editor {
-  background-color: #222;
-  color: white;
-}
-
-.dark-mode .ui.modal {
-    background-color: #1e1e1e;
-    color: #ffffff;
-}
-
-.dark-mode .ui.table {
-    background-color: #282828;
-    color: white;
-}
-
-.dark-mode .ui.input > input {
-    background-color: #333;
-    color: white;
-}
-
-body.dark-mode .ui.modal .header,
-body.dark-mode .ui.modal .content,
-body.dark-mode .ui.modal .actions {
-  background-color: inherit;
-  color: #e0e0e0;
-}
-
-body.dark-mode .ui.modal .button {
-  background-color: #424242;
-  color: #fff;
-  border: 1px solid #fff;
-}
-
-body.dark-mode .ui.modal .button:hover {
-  background-color: #333;
-}
-
-body.dark-mode .ui.segment.stackable.padded.grid {
-  background-color: #1e1e1e;
-  color: #e0e0e0;
-  border-color: #333;
-}
-
-body.dark-mode .ui.container.stackable.grid {
-  background-color: #1e1e1e;
-  color: #e0e0e0;
-}
-
-body.dark-mode .ui.modal .actions {
-  background-color: #1e1e1e;
-  border-top: 1px solid #333;
-}
-
-body.dark-mode .ui.segment .column {
-  color: #e0e0e0;
-}
-
-body.dark-mode .ui.segment .ui.small.header {
-  color: #ffffff;
-}
-
-body.dark-mode .ui.segment p {
-  color: #e0e0e0;
-}
-
-body.dark-mode .ui.segment .column h3,
-body.dark-mode .ui.segment .column h4,
-body.dark-mode .ui.segment .column h5,
-body.dark-mode .ui.segment .column p {
-  color: #e0e0e0;
-}
-
-body.dark-mode .ui.modal label,
-body.dark-mode .ui.form .field > label {
-    color: #e0e0e0;
-}
-
-/* Dark mode styles for accordion */
-body.dark-mode .ui.accordion {
-  background-color: #1e1e1e;
-  color: #e0e0e0;
-  border: 1px solid #333;
-}
-
-body.dark-mode .ui.accordion .title {
-  background-color: #222;
-  color: #ffffff;
-}
-
-body.dark-mode .ui.accordion .title:hover {
-  background-color: #333;
-  color: white;
-}
-
-body.dark-mode .ui.accordion .content {
-  background-color: #1e1e1e;
-  color: #e0e0e0;
-}
-
-body.dark-mode .ui.accordion .active.title {
-  background-color: #333;
-  color: #ffffff;
-}
-
-body.dark-mode .accordion-button-group .title:hover {
-  background-color: #222;
-  color: #ffffff;
-}
-
-body.dark-mode .ui.ui.styled.accordion {
-  background-color: #1e1e1e;
-  color: #e0e0e0;
-  border: 1px solid #333;
-}
-
-body.dark-mode .ui.styled.accordion .accordion .title {
-  background-color: #222;
-  color: #ffffff;
-}
-
-body.dark-mode .ui.styled.accordion .accordion .title:hover {
-  background-color: #333;
-  color: white;
-}
-
-body.dark-mode .ui.styled.accordion .accordion .content {
-  background-color: #1e1e1e;
-  color: #e0e0e0;
-}
-
-body.dark-mode .ui.styled.accordion .accordion .active.title {
-  background-color: #333;
-  color: #ffffff;
-}
-
-
-
-/* Dark mode styles for pagination */
-body.dark-mode .ui.pagination.menu {
-  background-color: #333;
-  border: 1px solid #555;
-}
-
-body.dark-mode .ui.pagination.menu .item {
-  background-color: #444;
-  color: #e0e0e0;
-  margin: 0;
-  display: inline-flex;
-  vertical-align: middle;
-}
-
-body.dark-mode .ui.pagination.menu .item:hover {
-  background-color: #555;
-}
-
-body.dark-mode .ui.pagination.menu .active.item {
-  background-color: #666;
-  color: white;
-  font-weight: bold;
-}
-
-
-
-body.dark-mode .ui.message {
-    background-color: #282828 ; 
-    color: #e0e0e0 !important;
-    border: 1px solid #444 !important; 
-    box-shadow: none !important; 
-}
