+++ ui/src/components/Tabs/AdminTab/UserEditor/UserEditorUserGroups.js
@@ -220,12 +220,13 @@ export default function UserEditorUserGroups(props) {
           key={semesterId}
           fluid
           styled
-
+          defaultActiveIndex={active ? 0 : false}
           panels={[
             {
               key: "StudentsTab-semester-selector-" + semesterId,
-              title: `${semesterMap[semesterId]["name"]} (${Object.keys(groupings["semesters"][semesterId])?.length})`,
-              active: active,
+              title: `${semesterMap[semesterId]["name"]} (${Object.keys(
+                groupings["semesters"][semesterId]
+              )?.length})`,
               content: {
                 content: createSemesterAccordion(
                   groupings["semesters"][semesterId],
