+++ ui/src/components/Tabs/ProjectsTab/Proposals.js
@@ -13,10 +13,12 @@ import ProjectEditorModal from "./ProjectEditorModal";
 import ProjectArchivePanel from "./ProjectArchivePanel";
 import _ from "lodash";
 import { config, PROJECT_STATUSES } from "../../util/functions/constants";
-import "../../../css/dashboard-proposal.css";
 import ProjectViewerModal from "./ProjectViewerModal";
 import WebsiteViewerModal from "./WebsiteViewerModal";
 import { isSemesterActive } from "../../util/functions/utils";
+import "./../../../css/components/proposal.css";
+import "./../../../css/containers/accordion.css";
+import "./../../../css/components/modal.css";
 
 const COLUMNS = {
   SEMESTER: "semester",
