+++ ui/src/components/util/components/Calendar.js
@@ -1,9 +1,9 @@
 import { useState, useEffect } from "react";
 import ToolTip from "../../Tabs/DashboardTab/TimelinesView/Timeline/ToolTip.js";
 import _ from "lodash";
-import "../../../css/calendar.css";
 import { max } from "moment";
 import { Popup } from "semantic-ui-react";
+import "./../../../css/components/calendar.css";
 
 const SPECIAL_DATES = {
   "01-01": "New Year's Day",
