+++ ui/src/css/components/announcement.css
@@ -1,0 +1,47 @@
+@import './../base/variables.css';
+
+    .announcement-container {
+        width: 100%;
+        display: flex;
+        flex-direction: column;
+        gap: 0.25em;
+    }
+
+    .announcement {
+        width: 100%;
+        background-color: #fff6f6;
+        border-radius: 0.5em;
+        padding: 0.5em;
+        margin-bottom: 0.25em;
+    }
+
+    .announcement .announcement-header {
+        display: flex;
+        margin-bottom: 0.5em;
+    }
+
+    .announcement .announcement-header :first-child {
+        flex-grow: 1;
+        padding-right: 0.25em;
+    }
+
+    .announcement .announcement-header * {
+        margin: 0;
+    }
+
+    .announcement-container-wrapper {
+        background-color: white;
+        padding: 10px;
+        border-radius: 5px;
+    }
+
+    /* Dark mode styles for Announcement */
+    body.dark-mode .announcement-container-wrapper {
+        background-color: #1e1e1e;
+        color: #e0e0e0;
+    }
+
+    body.dark-mode .announcement {
+        background-color: #333;
+        color: #e0e0e0;
+    }
