+++ ui/src/css/containers/footer.css
@@ -1,3 +1,5 @@
+@import './../base/variables.css';
+
 #footer {
     min-height: var(--footer-height);
     background-color: white;
