+++ ui/src/css/containers/header.css
@@ -1,3 +1,5 @@
+@import './../base/variables.css';
+
 #header {
     height: var(--header-height);
     padding-top: 2em;
