--- ui/src/css/dashboard-proposal.css
@@ -1,90 +1,0 @@
-.proposal-row-UNHANDLED-CASE {
-  background: blue;
-  color: red;
-}
-
-.proposal-row-UNHANDLED-CASE::before {
-  color: red;
-  content: "Something's not right...You're not supposed to see this.";
-}
-
-.proposal-row-gray {
-  background: #dddddd;
-  color: #666666;
-}
-
-.proposal-row-red {
-  background: #fff6f6;
-  color: black;
-}
-
-.proposal-row-yellow {
-  background: #fffaf3;
-  color: black;
-}
-
-.proposal-row-green {
-  background: #fcfff5;
-  color: black;
-}
-
-.proposal-row-blue {
-  background: #fcfff5;
-  color: black;
-}
-
-.proposal-row-purple {
-  background: #fcfff5;
-  color: black;
-}
-
-.proposal-status-action {
-  display: flex;
-  justify-content: space-between;
-}
-
-.proposal-status-action .dropdown {
-  margin-right: 0.33em;
-}
-
-td.attachments {
-  max-width: 25em;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-/* Dark Mode Styles */
-.dark-mode .proposal-row-UNHANDLED-CASE {
-    background: darkblue;       
-    color: #ff8080;             
-}
-
-.dark-mode .proposal-row-UNHANDLED-CASE::before {
-    color: #ff8080;
-}
-
-.dark-mode .proposal-row-gray {
-    background: #444444;      
-    color: #cccccc;            
-}
-
-.dark-mode .proposal-row-red {
-    background: #4d1c1c;        
-    color: #ffffff;             
-}
-
-.dark-mode .proposal-row-yellow {
-    background: #665c40;        
-    color: #ffffff;            
-}
-
-.dark-mode .proposal-row-green {
-    background: #2c3e2c;         
-    color: #e0f0e0;            
-}
-
-.dark-mode .proposal-row-blue {
-    background: #2d3a50;       
-    color: #d0e0f0;            
-}
