--- ui/src/css/proposal.css
@@ -1,23 +1,0 @@
-.proposal-submit-buttons {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.project-editor-input-container {
-  display: flex;
-  flex-direction: column;
-}
-
-.edited-input input,
-.edited-input.dropdown {
-  background-color: #fffaf3 !important;
-  color: #573a08 !important;
-}
-
-/* Dark Mode Styles */
-.dark-mode .edited-input input,
-.dark-mode .edited-input.dropdown {
-    background-color: #333 !important;
-    color: #e0e0e0 !important;
-}
