+++ ui/src/css/utils/responsive.css
@@ -1,0 +1,1 @@
+@import './../base/variables.css';
